reorganize commerce types/ folder
https://coveord.atlassian.net/browse/KIT-4635
