+++ packages/headless/src/ssr/commerce/controllers/filter-suggestions/headless-filter-suggestions-generator.ssr.ts
@@ -12,7 +12,7 @@ import {
   type FilterSuggestionsGenerator,
   type GeneratedFilterSuggestionsControllers,
 } from '../../../../controllers/commerce/filter-suggestions/headless-filter-suggestions-generator.js';
-import type {NonRecommendationControllerDefinitionWithoutProps} from '../../types/controllers.js';
+import type {NonRecommendationControllerDefinitionWithoutProps} from '../../types/controller-definitions.js';
 export type {
   FilterSuggestionsGenerator,
   GeneratedFilterSuggestionsControllers,
