+++ packages/headless/src/ssr/commerce/controllers/recent-queries-list/headless-recent-queries-list.ssr.ts
@@ -6,7 +6,7 @@ import {
   type RecentQueriesListProps,
   type RecentQueriesState,
 } from '../../../../controllers/commerce/recent-queries-list/headless-recent-queries-list.js';
-import type {NonRecommendationControllerDefinitionWithoutProps} from '../../types/controllers.js';
+import type {NonRecommendationControllerDefinitionWithoutProps} from '../../types/controller-definitions.js';
 
 export type {
   RecentQueriesListInitialState,
