+++ packages/headless/src/ssr/commerce/controllers/sort/headless-core-commerce-sort.ssr.ts
@@ -6,11 +6,11 @@ import type {
 import {buildProductListing} from '../../../../controllers/commerce/product-listing/headless-product-listing.js';
 import {buildSearch} from '../../../../controllers/commerce/search/headless-search.js';
 import {ensureAtLeastOneSolutionType} from '../../../../ssr/commerce/controller-utils.js';
-import {
-  type ControllerDefinitionOption,
-  SolutionType,
-  type SubControllerDefinitionWithoutProps,
-} from '../../types/controllers.js';
+import {SolutionType} from '../../types/controller-constants.js';
+import type {
+  ControllerDefinitionOption,
+  SubControllerDefinitionWithoutProps,
+} from '../../types/controller-definitions.js';
 
 export type {Sort, SortProps, SortState};
