+++ packages/headless/src/ssr/commerce/controllers/triggers/headless-query-trigger.ssr.ts
@@ -1,6 +1,6 @@
 import {buildQueryTrigger} from '../../../../controllers/commerce/triggers/headless-query-trigger.js';
 import type {QueryTrigger} from '../../../../controllers/core/triggers/headless-core-query-trigger.js';
-import type {SearchOnlyControllerDefinitionWithoutProps} from '../../types/controllers.js';
+import type {SearchOnlyControllerDefinitionWithoutProps} from '../../types/controller-definitions.js';
 
 export type {QueryTriggerState} from '../../../../controllers/core/triggers/headless-core-query-trigger.js';
 export type {QueryTrigger};
