+++ packages/headless/src/ssr/commerce/factories/build-factory.test.ts
@@ -5,7 +5,7 @@ import {getSampleCommerceEngineConfiguration} from '../../../app/commerce-engine
 import {buildLogger} from '../../../app/logger.js';
 import {defineProductList} from '../controllers/product-list/headless-product-list.ssr.js';
 import {defineSearchBox} from '../controllers/search-box/headless-search-box.ssr.js';
-import {SolutionType} from '../types/controllers.js';
+import {SolutionType} from '../types/controller-constants.js';
 import {buildFactory} from './build-factory.js';
 import {buildMockNavigatorContextProvider} from '../../../test/mock-navigator-context-provider.js';
 import {defineCart} from '../controllers/cart/headless-cart.ssr.js';
