+++ packages/headless/src/ssr/commerce/types/build.ts
@@ -8,17 +8,18 @@ import type {
   OptionsTuple,
 } from '../../common/types/utilities.js';
 import type {SSRCommerceEngine} from '../factories/build-factory.js';
+import type {SolutionType} from './controller-constants.js';
 import type {
   ControllerDefinitionsMap,
   EngineDefinitionBuildResult,
   EngineDefinitionControllersPropsOption,
-  SolutionType,
-} from './controllers.js';
+} from './controller-definitions.js';
 
 export interface BuildOptions<TEngineOptions> {
   extend?: OptionsExtender<TEngineOptions>;
 }
 
+// TODO: KIT-4610: Remove this type
 export type Build<
   TEngineOptions,
   TControllersMap extends ControllersMap,
