+++ packages/headless/src/ssr/commerce/types/controller-constants.ts
@@ -1,0 +1,12 @@
+export enum SolutionType {
+  search = 'search',
+  listing = 'listing',
+  standalone = 'standalone',
+  recommendation = 'recommendation',
+}
+
+const recommendationOptionKey = 'recommendation-internal-options';
+
+export const recommendationInternalOptionKey = Symbol.for(
+  recommendationOptionKey
+);
